Refactor signup flow into multi-step pages
